save if we are connected to the clipboard and reload in that state, also accpect the command line -connect and -disconnect for initial clipboard state
git-svn-id: svn://svn.code.sf.net/p/ditto-cp/code/trunk@481 595ec19a-5cb4-439b-94a8-42fb3063c22c

+++ Clip.cpp
@@ -681,7 +681,7 @@ bool CClip::AddToDataTable()
 			stmt.execDML();
 			stmt.reset();
 
-			pCF->m_lDBID = theApp.m_db.lastRowId();
+			pCF->m_lDBID = (long)theApp.m_db.lastRowId();
 		}
 	}
 	CATCH_SQLITE_EXCEPTION_AND_RETURN(false)
